Bug 1267989 part 3.  Make suppressing event handling on a document also block script execution for that document.  r=smaug
Note that we do this here instead of when we suspend timeouts, because it makes
it simpler to handle situations in which a window's document changes while
things are suspended/blocked.

The nsDocumentViewer.cpp is not strictly needed, but avoids some extraneous
calls to SuppressEventHandling with a 0 suppression count.

UltraBlame original commit: 81caeffdab2fa7027ccb873dfa48e99707f61344