Bug 1269924 - Move telemetry upload to NetworkUtils.isConnected. r=grisha
The concept of "background data" (as it exists in the Android options menu)
doesn't exist in the Android APIs - I think it should be covered under
isConnected. Thus, I removed our `isBackgroundDataEnabled` method.

One other network consideration, however: we may want to consider stopping
uploads on roaming.

In the previous implementation, we did not queue the ping for upload if
the network was not connected (in order to conserve disk space). However,
this doesn't allow us to see all of the days the user interacted with
the device (e.g. for engagement) so in this implementation, we always
queue the ping and stop in the UploadService if we're not connected.

MozReview-Commit-ID: 1mjnHq3l7Jj

UltraBlame original commit: 87a6b7a4960628f43aab39f76f31deb12f28ceac