Bug 1272578 part.1 Rename key name VolumeDown to AudioVolumeDown except on B2G r=smaug
VolumeDown was renamed to AudioVolumeDown in the latest draft and Chromium uses the new name. Therefore, we need to update this but Gaia uses the old name. So, we shouldn't rename on B2G until Gaia is fixed.

Note that this patch changes tests but they are not used by B2G. Therefore, just replacing with new name is enough.

Only forms.js is necessary #ifdef because the main purpose of forms.js is for B2G's IME framework. However, it's available on the other platforms if chrome needs to use it.

MozReview-Commit-ID: cq98qJnS8M

UltraBlame original commit: d0604395964c8aa9886755d53128bcfe5f3b2af8