Bug 1275755 - Remove MOZ_DUMP_ATOM_LEAKS. r=froydnj
I don't think anyone is using this anymore. It would be good to assert that there
are no leaks, but that doesn't pass for me in a local build, and I don't have time
to chase it.

UltraBlame original commit: 5bab999902fc0147cd24d3f5e1d7465f8ccd498e