Bug 1275604 - Make the touch-action pref live. r=botond
As part of making the pref live, we need to ensure that if the pref is flipped
while an input block is in the input queue, the behaviour that results is
reasonable. In particular, if a touch block is created while touch-action is
disabled, but the pref is enabled while that block is in the queue, it may
suddenly go from IsReadyForHandling() returning true to IsReadyForHandling()
returning false. This can result in the block getting stuck in the queue and
preventing future touch blocks from being processed at all. To handle this
case, we explicitly set the mAllowedTouchBehaviorSet flag to true if the block
is created with touch-action disabled. This prevents IsReadyForHandling() from
flipping to false in these cases.

MozReview-Commit-ID: CJq9NtMF1Bq

UltraBlame original commit: ece66ce5876bcdb00e7d5309fcc4b848390c9739