Bug 1265395 - Implement new low pass filter equation; r=padenot
If run against the version of the test using the old equations, 18 of the
26460 values would be larger than the maximum difference threshold, so I
updated the test code to use the new equations as well.

I also had to make a small increase to the maximum allowable glitch for
the tail test.

MozReview-Commit-ID: LrB3HufFWpJ

UltraBlame original commit: fed024a1b09241ef2dec9744bd001164247cd1b0