Bug 1280998 - Make a copy of the SurfaceDescriptor before sending it to IPC. r=nical
The IPDL code unconditionally calls forget() on any Shmem instances that are
sent over the IPC channel. This means that if the child process has a
SurfaceDescriptor containing a Shmem (such as a shmem-type SurfaceDescriptorBuffer)
then the shmem object in it will be zeroed out after sending it over IPC. In
order to still have access to the underlying SharedMemory, we need to make a
copy of the shmem or SurfaceDescriptor before doing the IPC call. Note that this
is true for safe and unsafe shmems.

MozReview-Commit-ID: KjEhPNiQhf9

UltraBlame original commit: 54d410626a98bfb517445a727c2683e81fa3105e