Bug 1266456 - part6: HTMLTooltip show() now accepts x,y offsets;r=bgrins
The autocomplete popup is displayed relatively to an anchor, but will be
shifted on the x-axis in order to be aligned with the word being completed.

To support this, the HTMLTooltip show() method now accepts x and y offsets.

MozReview-Commit-ID: 1cR3XFPdcVy

UltraBlame original commit: 88e3339aaec1cedbcc536836bd2fdea5b3f449c8