Bug 1283827 - Flatten transform-style:preserve-3d when combined with opacity. r=dbaron
This matches the behaviour required by the latest editor's draft spec and matches what blink is planning on shipping soon.
It may not be an entirely web compatible change, but we expect only a small number of websites to be affected. See
the bug for more details.

UltraBlame original commit: eac0c056235ee1174fc092fe159d2d5710331fbd